+++ src/test/java/com/mm/toy/service/BoardServiceTest.java
@@ -1,11 +1,11 @@
 package com.mm.toy.service;
 
-import com.mm.toy.domain.board.Board;
-import com.mm.toy.domain.board.BoardRepository;
-import com.mm.toy.domain.board.BoardRequestDto;
-import com.mm.toy.domain.user.User;
-import com.mm.toy.domain.user.UserRegisterDto;
-import com.mm.toy.domain.user.UserRepository;
+import com.mm.toy.domain.board.Entity.Board;
+import com.mm.toy.domain.board.Repository.BoardRepository;
+import com.mm.toy.domain.board.Dto.BoardRequestDto;
+import com.mm.toy.domain.user.Entity.User;
+import com.mm.toy.domain.user.Dto.UserRegisterDto;
+import com.mm.toy.domain.user.Repository.UserRepository;
 import com.mm.toy.global.service.DatabaseCleanup;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
